requirements.txt
altgraph==0.16.1
aniso8601==6.0.0
asn1crypto==0.24.0
astroid==2.0.4
backcall==0.1.0
bleach==2.1.3
certifi==2018.4.16
cffi==1.11.5
Click==7.0
cloudpickle==0.5.3
colorama==0.3.9
cryptography==2.2.1
cx-Freeze==5.1.1
cycler==0.10.0
dask==0.18.0
decorator==4.3.0
entrypoints==0.2.3
eventlet==0.23.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
flask-restful-swagger==0.20.1
flask-restplus==0.12.1
Flask-SocketIO==3.0.0
future==0.16.0
greenlet==0.4.13
h5py==2.8.0
html5lib==1.0.1
idna==2.7
imageio==2.1.2
imutils==0.5.3
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==5.1.5
isort==4.3.4
itsdangerous==1.1.0
jedi==0.12.0
Jinja2==2.10.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==1.2.1
lazy-object-proxy==1.3.1
macholib==1.11
MarkupSafe==1.1.1
matplotlib==2.0.0
mccabe==0.6.1
mistune==0.8.3
moviepy==0.2.3.5
msgpack==0.5.6
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.5.0
np-utils==0.5.5.0
numpy==1.11.1
olefile==0.45.1
opencv-python==3.1.0.0
pandas==0.23.1
pandocfilters==1.4.2
parso==0.2.1
patsy==0.5.0
pefile==2019.4.18
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.6.0
py2exe==0.9.2.2
pycparser==2.18
Pygments==2.2.0
PyInstaller==3.4
pylint==2.1.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyreadline==2.1
pytesseract==0.3.0
python-dateutil==2.7.3
python-engineio==2.1.1
python-socketio==1.9.0
pytz==2018.4
PyWavelets==0.5.2
pywin32==224
pywin32-ctypes==0.2.0
pywinpty==0.5.4
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
statsmodels==0.9.0
tensorflow-gpu==0.12.1
terminado==0.8.1
testpath==0.3.1
Theano==1.0.2
toolz==0.9.0
tornado==5.0.2
tqdm==4.23.4
traitlets==4.3.2
typed-ast==1.1.0
virtualenv==16.6.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==1.2.3
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
